+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Ignore protocol handler without support for availability check for file check schedule
 - Use user object instead of get_user() for better compatibility with older WordPress
+- Optimized usage of capability checks
 
 ## [5.2.1] - 19.05.2026